Refactor data classes and introduce database table objects for articles and authors
